Tuberculosis dataset

A dataset for the meta-analysis of the effect of BCG vaccine on incidence of tuberculosis.

Tuberculosis

Format

A data frame with 11 rows and 5 variables:

study

Places where the studies took place

authors

Authors of the studies

years

Year in which the study started

Risk

Value of the risk reported from the study

lower_ci

Value of the lower confidence interval of the risk reported from the study

upper_ci

Value of the upper confidence interval of the risk reported from the study

...

References

Sterne JA (2009) Meta-analysis in Stata: an updated collection from the Stata Journal. StataCorp LP, College Station, Texas