Sensitivity analysis for excess relative risk (ERR) or excess odds ratio (EOR) estimates meta-analysis

Fixed or Random effect model with either the standard approach or the alternative one could be used for the sensitivity analysis computation. The risk estimate could be excess relative risk (ERR) or excess odds ratio (EOR).

In the sensitivity analysis, each individual study is removed one at a time and the summarized estimate is computed to access the effect of the removed study on the overall pooled estimate.

exsens(
  study,
  err,
  u,
  l,
  d = NULL,
  conf.level = 0.95,
  test = c("FIXED", "RANDOM"),
  model = c("standard", "alternative")
)

Arguments

study
A vector (or column for dataframe, matrix) specifying the column reporting the author's name or the individual study's name

err
A numeric vector of the risk estimated from the individual studies

u
A numeric vector of the upper bound of the confidence interval of the risk reported from the individual studies.

l
A numeric vector of the lower bound of the confidence interval of the risk reported from the individual studies.

d
A numeric vector of the maximum dose reported from the individual studies.

conf.level
Coverage for confidence interval

test
Logical, indicating the statistical method to be used. The user have the choice between "FIXED" for the fixed effect model, and "RANDOM" for the random effect model.

model
Logical, indicating which statistical model should be used. The user have the choice between "standard" for the standard approach, and alternative" for the alternative approach for combining the risk estimate.

Value

Object of class "data.frame" that print the output from the exsens function. The following could be found from the output :

Study: Indication of the study removed from the pooled effect estimate

Effect: The pooled effect from the individual studies' estimate (ERR or EOR)

SE(Effect): The standard error of the pooled effect (see reference Richardson et al 2020 for more details)

Lower CI: The lower confidence interval bound of the pooled effect

Upper CI: The upper confidence interval bound of the pooled effect

Author

Kossi Abalo

Examples

study <- c("Canada", "Northern USA", "Chicago", "Georgia","Puerto", "Comm",
"Madanapalle", "UK", "South Africa", "Haiti", "Madras")
Risk <- c(0.205, 0.411, 0.254, 1.562, 0.712, 0.983, 0.804, 0.237, 0.625,
 0.198, 1.012)
lower_ci <- c(0.086, 0.134, 0.149, 0.374, 0.573, 0.582, 0.516, 0.179,
0.393, 0.078, 0.895)
upper_ci <- c(0.486, 1.257, 0.431, 6.528, 0.886, 1.659, 1.254, 0.312,
0.996, 0.499, 1.145)
dose <- c(32.586, 15.257, 72.431, 6.528, 10.886, 11.659, 17.254, 20.312,
10.996, 30.499, 41.145)

donne <- data.frame(cbind(study, Risk, lower_ci, upper_ci, dose))

donne$Risk <- as.numeric(as.character(donne$Risk))
donne$upper_ci <- as.numeric(as.character(donne$upper_ci))
donne$lower_ci <- as.numeric(as.character(donne$lower_ci))
donne$dose <- as.numeric(as.character(donne$dose))

exsens(study=donne$study, err=donne$Risk, u=donne$upper_ci,
l=donne$lower_ci, test = "FIXED", model = "standard")

#> STANDARD FIXED EFFECT MODEL EXCESS RISK ESTIMATE

#>           Study Effect SE(Effect) Lower CI Upper CI
#> 1        Madras   0.31       0.03     0.26     0.37
#> 2         Haiti   0.42       0.02     0.38     0.47
#> 3  South Africa   0.41       0.02     0.36     0.46
#> 4            UK   0.59       0.03     0.52     0.65
#> 5   Madanapalle   0.41       0.02     0.36     0.45
#> 6          Comm   0.41       0.02     0.36     0.46
#> 7        Puerto   0.38       0.03     0.33     0.43
#> 8       Georgia   0.41       0.02     0.37     0.46
#> 9       Chicago   0.43       0.03     0.38     0.48
#> 10 Northern USA   0.41       0.02     0.37     0.46
#> 11       Canada   0.43       0.02     0.38     0.47


exsens(study=donne$study, err=donne$Risk, u=donne$upper_ci,
l=donne$lower_ci, test = "RANDOM", model = "standard")

#> STANDARD RANDOM EFFECT MODEL EXCESS RISK ESTIMATE

#>           Study Effect SE(Effect) Lower CI Upper CI
#> 1        Madras   0.44       0.08     0.28     0.61
#> 2         Haiti   0.57       0.13     0.32     0.82
#> 3  South Africa   0.52       0.13     0.28     0.77
#> 4            UK   0.57       0.13     0.32     0.83
#> 5   Madanapalle   0.51       0.12     0.27     0.75
#> 6          Comm   0.50       0.12     0.26     0.73
#> 7        Puerto   0.51       0.13     0.26     0.76
#> 8       Georgia   0.53       0.12     0.30     0.76
#> 9       Chicago   0.57       0.13     0.31     0.83
#> 10 Northern USA   0.54       0.12     0.30     0.78
#> 11       Canada   0.57       0.13     0.32     0.82


exsens(study=donne$study, err=donne$Risk, u=donne$upper_ci,
l=donne$lower_ci, d=donne$dose, test = "FIXED",
 model = "alternative")

#> ALTERNATIVE FIXED EFFECT MODEL EXCESS RISK ESTIMATE

#>           Study Effect SE(Effect) Lower CI Upper CI
#> 1        Madras   0.43       0.05     0.38     0.49
#> 2         Haiti   0.66       0.04     0.60     0.72
#> 3  South Africa   0.65       0.04     0.59     0.71
#> 4            UK   0.79       0.04     0.72     0.87
#> 5   Madanapalle   0.64       0.04     0.59     0.70
#> 6          Comm   0.64       0.04     0.58     0.70
#> 7        Puerto   0.64       0.04     0.58     0.70
#> 8       Georgia   0.67       0.04     0.62     0.74
#> 9       Chicago   0.67       0.04     0.62     0.74
#> 10 Northern USA   0.65       0.04     0.59     0.71
#> 11       Canada   0.66       0.04     0.60     0.72


exsens(study=donne$study, err=donne$Risk, u=donne$upper_ci,
l=donne$lower_ci, d=donne$dose, test = "RANDOM",
 model = "alternative")

#> ALTERNATIVE RANDOM EFFECT MODEL EXCESS RISK ESTIMATE

#>           Study Effect SE(Effect) Lower CI Upper CI
#> 1        Madras   0.47       0.16     0.27     0.65
#> 2         Haiti   0.56       0.18     0.42     1.01
#> 3  South Africa   0.50       0.19     0.40     1.01
#> 4            UK   0.56       0.15     0.55     1.11
#> 5   Madanapalle   0.49       0.19     0.40     1.00
#> 6          Comm   0.48       0.19     0.40     0.99
#> 7        Puerto   0.50       0.21     0.37     1.04
#> 8       Georgia   0.48       0.19     0.43     1.03
#> 9       Chicago   0.55       0.18     0.43     1.03
#> 10 Northern USA   0.52       0.18     0.41     0.99
#> 11       Canada   0.55       0.18     0.42     1.01