Pooled risk estimate using the fixed effect model meta-analysis

Fixed effect model for standard meta-analysis of risk estimate (e.g relative risk (RR), odds ratio (OR) and hazard ratio (HR))

priskfix(rr, u, l, form = c("Log", "nonLog"), conf.level = 0.95)

Arguments

rr
A numeric vector of the risk estimated from the individual studies

u
A numeric vector of the upper bound of the confidence interval of the risk reported from the individual studies.

l
A numeric vector of the lower bound of the confidence interval of the risk reported from the individual studies.

form
Logical indicating the scale of the data. If Log, then the original data are in logarithmic scale.

conf.level
Coverage for confidence interval

Value

A list of a pooled result from the individual studies

Object of class "metaan.rr". A list that print the output from the priskfix function. The following could be found from the list :

rr_tot (Effect): The pooled effect from the individual studies' estimate (RR, OR, or HR)

sd_tot_lnRR (SE-Log(Effect)): The standard error of the pooled effect (see reference Richardson et al 2020 for more details)

l_tot (Lower CI): The lower confidence interval bound of the pooled effect (rr_tot)

u_tot (Upper CI): The upper confidence interval bound of the pooled effect (rr_tot)

Cochrane_stat (Cochran’s Q statistic): The value of the Cochrane's statistic of inter-study heterogeneity

Degree_freedom (Degree of Freedom): The degree of freedom

p_value (P-Value): The p-value of the statistic of Cochrane

I_square (Higgins’ and Thompson’s I^2 (%)): I square value in percent (%) indicating the amount of the inter-study heterogeneity

Examples

study <- c("Canada", "Northern USA", "Chicago", "Georgia","Puerto", "Comm",
 "Madanapalle", "UK", "South Africa", "Haiti", "Madras")
Risk <- c(0.205, 0.411, 0.254, 1.562, 0.712, 0.983, 0.804, 0.237, 0.625,
0.198, 1.012)
lower_ci <- c(0.086, 0.134, 0.149, 0.374, 0.573, 0.582, 0.516, 0.179, 0.393,
0.078, 0.895)
upper_ci <- c(0.486, 1.257, 0.431, 6.528, 0.886, 1.659, 1.254,
0.312, 0.996, 0.499, 1.145)

donne <- data.frame(cbind(study, Risk, lower_ci, upper_ci))

donne$Risk <- as.numeric(as.character(donne$Risk))
donne$upper_ci <- as.numeric(as.character(donne$upper_ci))
donne$lower_ci <- as.numeric(as.character(donne$lower_ci))

# on the log form
donne$ln_risk <- log(donne$Risk)
donne$ln_lower_ci <- log(donne$lower_ci)
donne$ln_upper_ci <- log(donne$upper_ci)


priskfix(rr=donne$Risk, u=donne$upper_ci, l=donne$lower_ci,
 form="nonLog", conf.level=0.95)

#>                                                   
#>   Standard meta-analysis with fixed effect model  
#> ------------------------------------------------- 
#>                                                   
#>   Effect SE-Log(Effect) Lower CI Upper CI
#>     0.73           0.05     0.67      0.8
#>                                                   
#> ------------------------------------------------- 
#>                                                   
#>             Test of heterogeneity : 
#>                                                   
#>   Cochran Q statistic Degree of Freedom P-Value
#>                125.05             10.00       0
#>                                                   
#> ------------------------------------------------- 
#>                                                   
#>   Higgins and Thompson I^2 (%)
#>                             92
#> _________________________________________________ 
#>                                                   



priskfix(rr=donne$ln_risk, u=donne$ln_upper_ci, l=donne$ln_lower_ci,
 form="Log", conf.level=0.95)

#>                                                   
#>   Standard meta-analysis with fixed effect model  
#> ------------------------------------------------- 
#>                                                   
#>   Effect SE-Log(Effect) Lower CI Upper CI
#>     0.73           0.05     0.67      0.8
#>                                                   
#> ------------------------------------------------- 
#>                                                   
#>             Test of heterogeneity : 
#>                                                   
#>   Cochran Q statistic Degree of Freedom P-Value
#>                125.05             10.00       0
#>                                                   
#> ------------------------------------------------- 
#>                                                   
#>   Higgins and Thompson I^2 (%)
#>                             92
#> _________________________________________________ 
#>