Sensitivity analysis for relative risk (or odds ratio) meta-analysis

Fixed or Random effect model could be used for the sensitivity analysis computation. The risk estimate could be e.g relative risk (RR), odds ratio (OR) or hazard ratio (HR).

In the sensitivity analysis, each individual study is removed one at a time and the summarized estimate is computed to access the effect of the removed study on the overall pooled estimate.

risksens(
  study,
  rr,
  u,
  l,
  form = c("Log", "nonLog"),
  test = c("FIXED", "RANDOM"),
  conf.level = 0.95
)

Arguments

study
A vector (or column for dataframe, matrix) specifying the column reporting the author's name or the individual study's name

rr
A numeric vector of the risk estimated from the individual studies

u
A numeric vector of the upper bound of the confidence interval of the risk estimated from the individual studies.

l
A numeric vector of the lower bound of the confidence interval of the risk estimated from the individual studies.

form
Logical, indicating the scale of the data. If Log, then the original data are in logarithmic scale.

test
Logical, indicating the statistical method to be used. The user have the choice between "FIXED" for the fixed effect model, and "RANDOM" for the random effect model.

conf.level
Coverage for confidence interval

Value

Object of class "data.frame" that print the output from the risksens function. The following could be found from the output :

Study: Indication of the study removed from the pooled effect estimate

Effect: The pooled effect from the individual studies' estimate (RR, OR, or HR)

SE-Log(Effect): The standard error of the pooled effect (see reference Richardson et al 2020 for more details)

Lower CI: The lower confidence interval bound of the pooled effect

Upper CI: The upper confidence interval bound of the pooled effect

Author

Kossi Abalo

Examples

study <- c("Canada", "Northern USA", "Chicago", "Georgia","Puerto", "Comm", "Madanapalle",
"UK", "South Africa", "Haiti", "Madras")
Risk <- c(0.205, 0.411, 0.254, 1.562, 0.712, 0.983, 0.804, 0.237, 0.625, 0.198, 1.012)
lower_ci <- c(0.086, 0.134, 0.149, 0.374, 0.573, 0.582, 0.516, 0.179, 0.393, 0.078, 0.895)
upper_ci <- c(0.486, 1.257, 0.431, 6.528, 0.886, 1.659, 1.254, 0.312, 0.996,
0.499, 1.145)

donne <- data.frame(cbind(study, Risk, lower_ci, upper_ci))

donne$Risk <- as.numeric(as.character(donne$Risk))
donne$upper_ci <- as.numeric(as.character(donne$upper_ci))
donne$lower_ci <- as.numeric(as.character(donne$lower_ci))

# on the log form
donne$ln_risk <- log(donne$Risk)
donne$ln_lower_ci <- log(donne$lower_ci)
donne$ln_upper_ci <- log(donne$upper_ci)

risksens(study=donne$study, rr=donne$ln_risk, u=donne$ln_upper_ci, l=donne$ln_lower_ci,
form="Log", test = "FIXED")

#> FIXED EFFECT MODEL RISK ESTIMATE ON LOG SCALE

#>           Study Effect SE Log(Effect) Lower CI Upper CI
#> 1        Madras   0.49           0.07     0.43     0.57
#> 2         Haiti   0.74           0.05     0.68     0.81
#> 3  South Africa   0.74           0.05     0.67     0.81
#> 4            UK   0.84           0.05     0.76     0.92
#> 5   Madanapalle   0.73           0.05     0.66     0.80
#> 6          Comm   0.72           0.05     0.66     0.80
#> 7        Puerto   0.74           0.05     0.67     0.81
#> 8       Georgia   0.73           0.05     0.67     0.80
#> 9       Chicago   0.76           0.05     0.69     0.83
#> 10 Northern USA   0.73           0.05     0.67     0.80
#> 11       Canada   0.74           0.05     0.68     0.81


risksens(study=donne$study, rr=donne$ln_risk, u=donne$ln_upper_ci, l=donne$ln_lower_ci,
form="Log", test = "RANDOM")

#> RANDOM EFFECT MODEL RISK ESTIMATE ON LOG SCALE

#>           Study Effect SE Log(Effect) Lower CI Upper CI
#> 1        Madras   0.46           0.22     0.30     0.71
#> 2         Haiti   0.55           0.21     0.36     0.83
#> 3  South Africa   0.50           0.23     0.32     0.78
#> 4            UK   0.58           0.17     0.41     0.81
#> 5   Madanapalle   0.48           0.23     0.31     0.76
#> 6          Comm   0.47           0.23     0.30     0.74
#> 7        Puerto   0.49           0.26     0.29     0.81
#> 8       Georgia   0.48           0.22     0.31     0.73
#> 9       Chicago   0.55           0.21     0.36     0.84
#> 10 Northern USA   0.52           0.22     0.34     0.79
#> 11       Canada   0.55           0.21     0.36     0.84


risksens(study=donne$study, rr=donne$Risk, u=donne$upper_ci, l=donne$lower_ci,
form="nonLog", test = "FIXED")

#> FIXED EFFECT MODEL RISK ESTIMATE ON NON-LOG SCALE

#>           Study Effect SE Log(Effect) Lower CI Upper CI
#> 1        Madras   0.49           0.07     0.43     0.57
#> 2         Haiti   0.74           0.05     0.68     0.81
#> 3  South Africa   0.74           0.05     0.67     0.81
#> 4            UK   0.84           0.05     0.76     0.92
#> 5   Madanapalle   0.73           0.05     0.66     0.80
#> 6          Comm   0.72           0.05     0.66     0.80
#> 7        Puerto   0.74           0.05     0.67     0.81
#> 8       Georgia   0.73           0.05     0.67     0.80
#> 9       Chicago   0.76           0.05     0.69     0.83
#> 10 Northern USA   0.73           0.05     0.67     0.80
#> 11       Canada   0.74           0.05     0.68     0.81


risksens(study=donne$study, rr=donne$Risk, u=donne$upper_ci, l=donne$lower_ci,
form="nonLog", test = "RANDOM")

#> RANDOM EFFECT MODEL RISK ESTIMATE ON NON-LOG SCALE

#>           Study Effect SE Log(Effect) Lower CI Upper CI
#> 1        Madras   0.46           0.22     0.30     0.71
#> 2         Haiti   0.55           0.21     0.36     0.83
#> 3  South Africa   0.50           0.23     0.32     0.78
#> 4            UK   0.58           0.17     0.41     0.81
#> 5   Madanapalle   0.48           0.23     0.31     0.76
#> 6          Comm   0.47           0.23     0.30     0.74
#> 7        Puerto   0.49           0.26     0.29     0.81
#> 8       Georgia   0.48           0.22     0.31     0.73
#> 9       Chicago   0.55           0.21     0.36     0.84
#> 10 Northern USA   0.52           0.22     0.34     0.79
#> 11       Canada   0.55           0.21     0.36     0.84